+++ docs/en/changes/changes.md
@@ -22,6 +22,7 @@
 * [Breaking Change] Add layer field to event, report an event without layer is not allowed.
 * Fix ES flush thread stops when flush schedule task throws exception, such as ElasticSearch flush failed.
 * Fix ES BulkProcessor in BatchProcessEsDAO was initialized multiple times and created multiple ES flush schedule tasks.
+* HttpServer support configure the allowed HTTP methods. 
 
 #### UI